requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
black==23.12.1
flake8==6.1.0
